% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegLobes-package.R
\docType{package}
\name{ggsegLobes-package}
\alias{ggsegLobes}
\alias{ggsegLobes-package}
\title{ggsegLobes: Lobar Parcellation Atlas for the 'ggsegverse' Ecosystem}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Lobar cortical parcellation (frontal, parietal, temporal, occipital, insula, per hemisphere) derived from the Desikan-Killiany atlas by dissolving its cortical regions into lobes. Provided as a 'ggseg_atlas' object for use with 'ggseg'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gbarisano/ggsegLobes}
  \item Report bugs at \url{https://github.com/gbarisano/ggsegLobes/issues}
}

}
\author{
\strong{Maintainer}: Giuseppe Barisano \email{barisano@stanford.edu} (\href{https://orcid.org/0000-0001-5598-1369}{ORCID})

}
\keyword{internal}
